add verification of the token in server side and inprove the token management in the cookie

+++ php/delNews.php
@@ -1,19 +1,25 @@
 <?php
+require_once('CONF.php');
+require_once('jwt_helper.php');
+
 $data = json_decode(file_get_contents("php://input"));
-$id = $data->id;
+$token = $data->token;
+$token = JWT::decode($token, $SECRET_KEY);
 
-$sql ="DELETE FROM news WHERE ID='$id'";
+if ($token->admin){
+  $data = json_decode(file_get_contents("php://input"));
+  $id = $data->id;
 
- /*INSERT INTO news (TITLE,BODY,AUTHOR,DATE_CREATE)
- VALUES ($POST['title'],$POST['body'],$POST['uthor'],$POST['bdaytime']);
-*/
-$database = new SQLite3('../sqlite/newsdb.db');
+  $sql ="DELETE FROM news WHERE ID='$id'";
 
-$ret = $database->exec($sql);
-if(!$ret){
-  echo $database->lastErrorMsg();
-} else {
-  echo "success";
-}
+  $database = new SQLite3('../sqlite/newsdb.db');
 
+  $ret = $database->exec($sql);
+  if(!$ret){
+    echo $database->lastErrorMsg();
+  } else {
+    echo "success";
+  }
+}else
+  echo "token is not valid";
 ?>
